Refactor navigation setup
